requirements.txt
Flask==3.1.0
pydantic-ai==0.0.55
mcp==1.6.0
python-dotenv==1.1.0
requests==2.32.3
langdetect==1.0.9
neo4j==5.8.0
graphiti==0.1.0
serpapi